Add files via upload
My group's poster that took part in the 2022 HEAD Competition at Mohawk. This dealt with a dataset dealing with Municipal parking Hamilton while looking at crime activity in the Hamilton area. Density of parking and a portion of users were analyzed to identify opportunities as well. This poster came in 1st in the 1st phase in the competition.